Add Pokemon component
This adds the initial redux struture to the
application.

the pokemon component should receve a selected
pokemon to display the pokemon info.

it has been added the models for the pokemon structure,
as the service to consume the pokeap/pokemon route.

it has been addied the jest test for all initial redux
and react components.

it also been add support for scss.